[BE] Feat #112: 비공개 버킷/후기에 타인은 리액션, 댓글 CRUD 불가

+++ backend/src/main/java/com/ggums/ggumtle/service/CommentBucketService.java
@@ -151,6 +151,11 @@ public class CommentBucketService {
             throw new CustomException(ExceptionType.NOT_VALID_USER);
         }
 
+        // 버킷이 비공개인데 user가 버킷의 주인이 아닌 경우 예외 처리
+        if (comment.getBucket().getIsPrivate() && !user.getId().equals(comment.getBucket().getUser().getId())) {
+            throw new CustomException(ExceptionType.BUCKET_NOT_VALID);
+        }
+
         commentBucketRepository.delete(comment);
 
         // user가 버킷 작성자(writer)를 팔로우하고 있는 경우 user -> writer 친밀도 감소
@@ -172,6 +177,11 @@ public class CommentBucketService {
         // 요청한 사용자와 댓글 작성자가 다르면 오류 반환
         if (!(comment.getUser().getId().equals(user.getId()))) {
             throw new CustomException(ExceptionType.NOT_VALID_USER);
+        }
+
+        // 버킷이 비공개인데 user가 버킷의 주인이 아닌 경우 예외 처리
+        if (comment.getBucket().getIsPrivate() && !user.getId().equals(comment.getBucket().getUser().getId())) {
+            throw new CustomException(ExceptionType.BUCKET_NOT_VALID);
         }
 
         comment.update(requestDto.getContext());
